refactor: split camera logic
